Replaced some tests in the code with json tests.
